Added View Model,Repository, DI Setup NetworkModule, Service Added Adapter, listItems, internet permission Added ViewBindingAdapter, extension function

+++ app/src/main/java/com/stocks/assignment/model/HoldingResponseModel.kt
@@ -20,6 +20,6 @@ data class HoldingResponseModel(
     val responseType: String? = null,
 
     @SerializedName("timestamp")
-    val timestamp: Int? = null
+    val timestamp: Long? = null
 
 )
